add printf

+++ newer/ProxyDASH/src/ClientManager.cpp
@@ -10,7 +10,7 @@
 #include "InterfacesManager.h"
 #include "StatManager.h"
 
-static long int timevaldiff_usec(struct timeval *start, struct timeval *end) {
+static long long timevaldiff_usec(struct timeval *start, struct timeval *end) {
 	return (end->tv_sec * 1000000 + end->tv_usec) - (start->tv_sec * 1000000 + start->tv_usec);
 }
